// types/survey.ts
export interface AjisSurveyListItem {
  id_survey: number;
  tgl_survey: string | null;
  petugas_survey: string | null;
  id_anak: string;
  nama_lengkap: string | null;
  jns_kel: string | null;
  jenjang_pendidikan: string | null;
  kantor_id: string | null;
  nama_kantor: string | null;
  id_wilayah_pembinaan: string | null;
  nama_wilayah: string | null;
  asnaf: string | null;
  hasil_kesimpulan_survey: string | null;
  asnaf_anak: 'yatim' | 'piatu' | 'dhuafa' | null;
  date_insert: string | null;
}